rename terrascript_client to terrascript_aws_client
