+++ pyproject.toml
@@ -6,6 +6,6 @@ extend-exclude = '''
 ^/(
     reconcile/terraform_resources.py
     | reconcile/utils/terraform_client.py
-    | reconcile/utils/terrascript_client.py
+    | reconcile/utils/terrascript_aws_client.py
 )$
 '''
