+++ reconcile/aws_iam_keys.py
@@ -6,7 +6,7 @@ from reconcile import queries
 
 from reconcile.utils.defer import defer
 from reconcile.utils.aws_api import AWSApi
-from reconcile.utils.terrascript_client import TerrascriptClient as Terrascript
+from reconcile.utils.terrascript_aws_client import TerrascriptClient as Terrascript
 
 QONTRACT_INTEGRATION = "aws-iam-keys"
