+++ setup.cfg
@@ -256,7 +256,7 @@ check_untyped_defs = False
 [mypy-reconcile.utils.terraform_client.*]
 check_untyped_defs = False
 
-[mypy-reconcile.utils.terrascript_client.*]
+[mypy-reconcile.utils.terrascript_aws_client.*]
 check_untyped_defs = False
 
 [mypy-tools.app_interface_reporter.*]
